NIL check for missing payment profile for subscriptions
I was trying to access the payment_profile of a subscription that did not have
one. When doing so an Undefined method "credit_card" exception was being thrown.
I added a check to the payment_profile method to check and see if the instance
responds to credit_card before attempting to return it.