Add Chargify::Allocation::Payment class
This is needed so that ActiveResource will not use Chargify::Payment
when there is a Payment inside an Allocation.

This Payment is an output-only attribute of an Allocation.

The Payment class should inherit from Base so that its initialize method is defined properly.

Fixes https://github.com/chargify/chargify_api_ares/issues/131